add:emotionインストール
