+++ app/.babelrc
@@ -1,0 +1,14 @@
+{
+    "presets": [
+      [
+        "next/babel",
+        {
+          "preset-react": {
+            "runtime": "automatic",
+            "importSource": "@emotion/react"
+          }
+        }
+      ]
+    ],
+    "plugins": ["@emotion/babel-plugin"]
+  }
