+++ app/package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/babel-plugin": "^11.7.2",
+    "@emotion/react": "^11.7.1",
     "axios": "^0.24.0",
     "firebase": "^9.2.0",
     "next": "^12.0.8-canary.1",
