+++ app/tsconfig.json
@@ -17,6 +17,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
+    "types": ["@emotion/react/types/css-prop"],
     "incremental": true
   },
   "include": [
